requirements.txt
Django==4.2
django~=4.0
django-extensions
django-heroku
dj-database-url
gunicorn
python-dotenv
pydotplus
psycopg2
psycopg2-binary
whitenoise
secret-key-generator
djangorestframework